pyproject.toml
[project]
name = "rag-with-clustered-documents"
version = "0.1.0"
description = ""
authors = [
    {name = "Maximilian S"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "datasets (>=4.0.0,<5.0.0)",
    "ollama (>=0.5.1,<0.6.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "transformers (>=4.54.1,<5.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "faiss-cpu (>=1.11.0.post1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
